+++ core/interfaces.go
@@ -2,6 +2,7 @@ package core
 
 import (
 	"core-building-block/core/model"
+	"core-building-block/driven/storage"
 
 	log "github.com/rokmetro/logging-library/loglib"
 )
@@ -11,22 +12,6 @@ type Services interface {
 	SerGetVersion(l *log.Log) string
 	SerGetAuthTest(l *log.Log) string
 	SerGetCommonTest(l *log.Log) string
-}
-
-type servicesImpl struct {
-	app *Application
-}
-
-func (s *servicesImpl) SerGetVersion(l *log.Log) string {
-	return s.app.serGetVersion(l)
-}
-
-func (s *servicesImpl) SerGetAuthTest(l *log.Log) string {
-	return s.app.serGetAuthTest(l)
-}
-
-func (s *servicesImpl) SerGetCommonTest(l *log.Log) string {
-	return s.app.serGetCommonTest(l)
 }
 
 //Administration exposes administration APIs for the driver adapters
@@ -39,34 +24,7 @@ type Administration interface {
 	AdmUpdateGlobalConfig(setting string) error
 
 	AdmCreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error)
-}
-
-type administrationImpl struct {
-	app *Application
-}
-
-func (s *administrationImpl) AdmGetTest() string {
-	return s.app.admGetTest()
-}
-
-func (s *administrationImpl) AdmGetTestModel() string {
-	return s.app.admGetTestModel()
-}
-
-func (s *administrationImpl) AdmCreateGlobalConfig(setting string) (*model.GlobalConfig, error) {
-	return s.app.admCreateGlobalConfig(setting)
-}
-
-func (s *administrationImpl) AdmGetGlobalConfig() (*model.GlobalConfig, error) {
-	return s.app.admGetGlobalConfig()
-}
-
-func (s *administrationImpl) AdmUpdateGlobalConfig(setting string) error {
-	return s.app.admUpdateGlobalConfig(setting)
-}
-
-func (s *administrationImpl) AdmCreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error) {
-	return s.app.admCreateOrganization(name, requestType, requiresOwnLogin, loginTypes, organizationDomains)
+	AdmUpdateOrganization(ID string, name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) error
 }
 
 //Encryption exposes APIs for the Encryption building block
@@ -74,44 +32,28 @@ type Encryption interface {
 	EncGetTest() string
 }
 
-type encryptionImpl struct {
-	app *Application
-}
-
-func (s *encryptionImpl) EncGetTest() string {
-	return s.app.encGetTest()
-}
-
 //BBs exposes users related APIs used by the platform building blocks
 type BBs interface {
 	BBsGetTest() string
 }
 
-type bbsImpl struct {
-	app *Application
-}
-
-func (s *bbsImpl) BBsGetTest() string {
-	return s.app.bbsGetTest()
-}
-
 //Storage is used by core to storage data - DB storage adapter, file storage adapter etc
 type Storage interface {
-	SetStorageListener(storageListener StorageListener)
+	RegisterStorageListener(storageListener storage.StorageListener)
 
 	CreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 	GetGlobalConfig() (*model.GlobalConfig, error)
 	SaveGlobalConfig(setting *model.GlobalConfig) error
 
 	CreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error)
+	UpdateOrganization(ID string, name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) error
 }
 
 //StorageListener listenes for change data storage events
-type StorageListener interface {
-}
 
-type storageListenerImpl struct {
-	app *Application
+type CoreStorageListener struct {
+	app *application
+	storage.DefaultStorageListenerImpl
 }
 
 //ApplicationListener represents application listener