+++ driven/storage/adapter.go
@@ -2,17 +2,19 @@ package storage
 
 import (
 	"context"
-	"core-building-block/core"
 	"core-building-block/core/model"
+	"errors"
 	"fmt"
-
 	"strconv"
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/rokmetro/auth-library/authservice"
 	log "github.com/rokmetro/logging-library/loglib"
 	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 type organization struct {
@@ -48,14 +50,43 @@ func (sa *Adapter) Start() error {
 	return err
 }
 
-//SetStorageListener sets listener for the storage
-func (sa *Adapter) SetStorageListener(storageListener core.StorageListener) {
-	sa.db.listener = storageListener
+//RegisterStorageListener registers a data change listener with the storage adapter
+func (sa *Adapter) RegisterStorageListener(storageListener StorageListener) {
+	sa.db.listeners = append(sa.db.listeners, storageListener)
 }
 
 //ReadTODO TODO TODO
 func (sa *Adapter) ReadTODO() error {
 	return nil
+}
+
+//FindAuthConfig finds the auth document from DB by orgID and appID
+func (sa *Adapter) FindAuthConfig(orgID string, appID string, authType string) (*model.AuthConfig, error) {
+	filter := bson.D{primitive.E{Key: "org_id", Value: orgID}, primitive.E{Key: "app_id", Value: appID}, primitive.E{Key: "type", Value: authType}}
+	var result *model.AuthConfig
+	err := sa.db.authConfigs.FindOne(filter, &result, nil)
+	if err != nil {
+		return nil, err
+	}
+	if result == nil {
+		return nil, fmt.Errorf("no auth config found for orgID %s, appID %s, authType %s", orgID, appID, authType)
+	}
+	return result, nil
+}
+
+//LoadAuthConfigs finds all auth config documents in the DB
+func (sa *Adapter) LoadAuthConfigs() (*[]model.AuthConfig, error) {
+	filter := bson.D{}
+	var result []model.AuthConfig
+	err := sa.db.authConfigs.Find(filter, &result, nil)
+	if err != nil {
+		return nil, err
+	}
+	if len(result) == 0 {
+		return nil, errors.New("no auth config documents found")
+	}
+
+	return &result, nil
 }
 
 //CreateGlobalConfig creates global config
@@ -111,8 +142,8 @@ func (sa *Adapter) SaveGlobalConfig(gc *model.GlobalConfig) error {
 
 		err = sessionContext.CommitTransaction(sessionContext)
 		if err != nil {
-			//TODO print
-			//log.Printf("error on commiting a transaction - %s", err)
+			abortTransaction(sessionContext)
+			fmt.Errorf("error on commiting a transaction - %s", err)
 			return err
 		}
 		return nil
@@ -147,6 +178,65 @@ func (sa *Adapter) CreateOrganization(name string, requestType string, requiresO
 	return &resOrg, nil
 }
 
+//UpdateOrganization updates an organization
+func (sa *Adapter) UpdateOrganization(ID string, name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) error {
+
+	now := time.Now()
+
+	updatOrganizationFilter := bson.D{primitive.E{Key: "_id", Value: ID}}
+	updateOrganization := bson.D{
+		primitive.E{Key: "$set", Value: bson.D{
+			primitive.E{Key: "name", Value: name},
+			primitive.E{Key: "type", Value: requestType},
+			primitive.E{Key: "requires_own_login", Value: requiresOwnLogin},
+			primitive.E{Key: "login_types", Value: loginTypes},
+			primitive.E{Key: "config.domains", Value: organizationDomains},
+			primitive.E{Key: "config.date_updated", Value: now},
+			primitive.E{Key: "date_updated", Value: now},
+		}},
+	}
+
+	result, err := sa.db.organizations.UpdateOne(updatOrganizationFilter, updateOrganization, nil)
+	if err != nil {
+		return err
+	}
+	if result.MatchedCount == 0 {
+		return errors.New("there is no organziation for the provided id")
+	}
+
+	return nil
+}
+
+//GetServiceRegs fetches the requested service registration records
+func (sa *Adapter) GetServiceRegs(serviceIDs []string) ([]authservice.ServiceReg, error) {
+	var filter bson.M
+	for _, serviceID := range serviceIDs {
+		if serviceID == "all" {
+			filter = bson.M{}
+			break
+		}
+	}
+	if filter == nil {
+		filter = bson.M{"service_id": bson.M{"$in": serviceIDs}}
+	}
+
+	var result []authservice.ServiceReg
+	err := sa.db.serviceRegs.Find(filter, &result, nil)
+	return result, err
+}
+
+//SaveServiceReg saves the service registration to the storage
+func (sa *Adapter) SaveServiceReg(reg *authservice.ServiceReg) error {
+	filter := bson.M{"service_id": reg.ServiceID}
+	opts := options.Replace().SetUpsert(true)
+	err := sa.db.serviceRegs.ReplaceOne(filter, reg, opts)
+	if err != nil {
+		return fmt.Errorf("error saving service reg for service id %s: %v", reg.ServiceID, err)
+	}
+
+	return nil
+}
+
 //NewStorageAdapter creates a new storage adapter instance
 func NewStorageAdapter(mongoDBAuth string, mongoDBName string, mongoTimeout string, logger *log.StandardLogger) *Adapter {
 	timeout, err := strconv.Atoi(mongoTimeout)
@@ -166,4 +256,14 @@ func abortTransaction(sessionContext mongo.SessionContext) {
 		//TODO - log
 	}
 
+}
+
+type StorageListener interface {
+	OnAuthConfigUpdated()
+}
+
+type DefaultStorageListenerImpl struct {
+}
+
+func (d *DefaultStorageListenerImpl) OnAuthConfigUpdated() {
 }